+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 
 * `add-or-replace-column(s)` replaced by `add-column(s)` (`add-or-replace-column(s)` is marked as deprecated)
 
+## [5.00-beta-28]
+
+### Added
+
+* `let-dataset` - to simulate `tibble` from R
+
+### Fixed
+
+* Adding a column to an empty dataset returned empty dataset
+
 ## [5.00-beta-27]
 
 ### Changed